moved methods to super class
